Cross compile containerd
Current distributions are not able to keep up with the pace that
projects like containerd are releasing, preventing users using
new features until a new distribution is released.

Containerd provides binaries in a tarball that are easy to use, but
they doesn't support multiple architectures.

This patch adds a script that allows to cross-compile and release
containerd for multiple archicectures.

Signed-off-by: Antonio Ojea <antonio.ojea.garcia@gmail.com>